files been uploaded.
detailed desc of the tranquilizer which is developed in HTML CSS and Javascript and then deployed in netlify. and then i switched over the web application to a android application using web to app website which is pretty useful to convert web applications to android applications. 
This is a freelance project which made me jumped into this. 
if you find any error in this kindly do mention and we will be work to rectify the error. 

Login details:
*** Username - admin  .  password - password ***